A simple-minded, RESTful login service
This program reads a .htaccess file and provides session cookies that
it stores in a directory.

This means that it's neither suited for many users (lookups are linear
in time), nor suited for many sessions (directories do get large).

There's no cleanup, either, so old sessions just don't get deleted.