Fix login redirection to not overwrite the first destination
The previous iteration of this code had a problem where a visit to /
would sometimes end up redirecting to /favicon.ico: Turns out browsers
make multiple requests in sequence.

This assumes (of course) that the first requested URL is the one we
want; I hope that's right, but if not we can solve this in nginx.conf
I suppose.