Fix session cleanup, add a test
Turns out I had a logic error in session cleanup logic: It would
remove only those sessions that had not yet expired. Oops.

Also rephrase the "ExpiredAt" test to be less confusing, and reduce
the time slack in cookie creation, so that cleanup becomes testable in
a more-trivial amount of time.